+++ include/rain/networking/smtp/command.hpp
@@ -41,7 +41,7 @@ namespace Rain::Networking::Smtp {
 
 		public:
 		// Maps strings to Value. Commands are case-agnostic.
-		inline static std::unordered_map<
+		static inline std::unordered_map<
 			std::string,
 			Value,
 			String::CaseAgnosticHash,